Move class AlertInfo to Sensor interface so it can be used by other sensors (#2)
Alert info is a generic class that could be reused across different
sensors. Therefore, this PR moves it from SuccessrateSensor to
the Sensor interafce.
